Move fd into sys
